# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Moling API",
    "description": "墨灵 — AI 小说创作平台后端服务",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/generation/{task_id}": {
      "get": {
        "tags": [
          "generation",
          "generation"
        ],
        "summary": "Get Task Status",
        "description": "Get generation task status.\n\nAuthentication is checked first via Depends(get_current_user).\nReturns 401 if not authenticated, 404 if task not found,\n403 if task belongs to another user.",
        "operationId": "get_task_status_api_v1_generation__task_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatusResp"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/generation/{task_id}/cancel": {
      "delete": {
        "tags": [
          "generation",
          "generation"
        ],
        "summary": "Cancel Task",
        "description": "Cancel a generation task.\n\nAuthentication is checked first via Depends(get_current_user).\nReturns 401 if not authenticated, 404 if task not found,\n403 if task belongs to another user.",
        "operationId": "cancel_task_api_v1_generation__task_id__cancel_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Cancel Task Api V1 Generation  Task Id  Cancel Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/generation/history": {
      "get": {
        "tags": [
          "generation",
          "generation"
        ],
        "summary": "Get Generation History",
        "description": "Get generation task history.",
        "operationId": "get_generation_history_api_v1_generation_history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Page size",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Page size"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Generation History Api V1 Generation History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/genre/prefill": {
      "post": {
        "tags": [
          "genre"
        ],
        "summary": "Trigger Prefill",
        "description": "B1-B5 完整冷启动流程。\n\n1. 验证用户对项目的所有权（TODO: 增加项目权限检查）\n2. 执行 B1-B5 冷启动\n3. 返回预填数据供前端展示",
        "operationId": "trigger_prefill_api_v1_genre_prefill_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrefillRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrefillResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/genre/prefill/{project_id}": {
      "get": {
        "tags": [
          "genre"
        ],
        "summary": "Get Prefill",
        "description": "获取预填结果供用户审核（B5 审核界面用）。",
        "operationId": "get_prefill_api_v1_genre_prefill__project_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrefillSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/genre/prefill/{project_id}/confirm": {
      "post": {
        "tags": [
          "genre"
        ],
        "summary": "Confirm Prefill",
        "description": "用户确认预填数据并入库（B5 最终步骤）。",
        "operationId": "confirm_prefill_api_v1_genre_prefill__project_id__confirm_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmRequest",
                "default": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfirmRequest": {
        "properties": {
          "modifications": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ModificationItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modifications"
          }
        },
        "type": "object",
        "title": "ConfirmRequest",
        "description": "确认预填请求（可选修改）"
      },
      "ConfirmResponse": {
        "properties": {
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "state": {
            "type": "string",
            "title": "State",
            "default": "confirmed"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "message"
        ],
        "title": "ConfirmResponse",
        "description": "确认入库响应"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModificationItem": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "action": {
            "type": "string",
            "title": "Action",
            "default": "update"
          },
          "value": {
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "path"
        ],
        "title": "ModificationItem",
        "description": "用户修改项"
      },
...
      "PrefillRequest": {
        "properties": {
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "genre": {
            "type": "string",
            "maxLength": 20,
            "minLength": 1,
            "title": "Genre"
          },
          "synopsis": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 10,
            "title": "Synopsis"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "genre",
          "synopsis"
        ],
        "title": "PrefillRequest",
        "description": "触发冷启动请求"
      },
      "PrefillResponse": {
        "properties": {
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "genre": {
            "type": "string",
            "title": "Genre"
          },
          "synopsis": {
            "type": "string",
            "title": "Synopsis"
          },
          "profile_source": {
            "type": "string",
            "title": "Profile Source"
          },
          "profile_version": {
            "type": "string",
            "title": "Profile Version"
          },
          "vault": {
            "additionalProperties": true,
            "type": "object",
            "title": "Vault"
          },
          "dynamic_layer": {
            "additionalProperties": true,
            "type": "object",
            "title": "Dynamic Layer"
          },
          "card_pool": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Card Pool"
          },
          "opening_directions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Opening Directions"
          },
          "async_analysis_triggered": {
            "type": "boolean",
            "title": "Async Analysis Triggered"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "genre",
          "synopsis",
          "profile_source",
          "profile_version",
          "vault",
          "dynamic_layer",
          "card_pool",
          "opening_directions",
          "async_analysis_triggered",
          "session_id"
        ],
        "title": "PrefillResponse",
        "description": "预填数据响应"
      },
      "PrefillSessionResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "genre": {
            "type": "string",
            "title": "Genre"
          },
          "synopsis": {
            "type": "string",
            "title": "Synopsis"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "prefill_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prefill Data"
          },
          "profile_source": {
            "type": "string",
            "title": "Profile Source"
          },
          "profile_version": {
            "type": "string",
            "title": "Profile Version"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "project_id",
          "genre",
          "synopsis",
          "state",
          "profile_source",
          "profile_version"
        ],
        "title": "PrefillSessionResponse",
        "description": "预填会话详情响应"
      },
...
      "TaskStatusResp": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id",
            "description": "生成任务 ID (UUID)"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "任务状态"
          },
          "progress_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress Stage",
            "description": "当前进度阶段"
          },
          "progress_percent": {
            "type": "integer",
            "title": "Progress Percent",
            "description": "进度百分比",
            "default": 0
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message",
            "description": "错误信息"
          },
          "output_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Data",
            "description": "输出数据"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "status"
        ],
        "title": "TaskStatusResp",
        "description": "Polling response for generation task progress."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}